(api-cat) Refactor separate ResponseAdapter class for processing elacticsearch response
